router/requirements.txt
fastmcp>=2.0.0
python-dotenv>=1.0.0
mcp-email-server>=0.1.0
